moved app.js to src folder - removed react-native stack navigator
