+++ nativeClient/package.json
@@ -9,8 +9,7 @@
 	"dependencies": {
 		"axios": "^0.17.0",
 		"react": "16.0.0",
-		"react-native": "0.50.2",
-		"react-navigation": "^1.0.0-beta.19"
+		"react-native": "0.50.2"
 	},
 	"devDependencies": {
 		"babel-jest": "21.2.0",
